refactor(android): extract shared line-JSON Conn into conn.rs
