+++ crates/glass-android/src/agent.rs
@@ -3,8 +3,6 @@
 //! `localabstract:glass-agent`. `AgentClient` is the request/response client; `AgentRegistry`
 //! owns the device server's lifecycle. Everything degrades to the adb paths on failure.
 
-use std::io::{BufRead, BufReader, Write};
-use std::net::TcpStream;
 use std::process::{Child, Command, Stdio};
 use std::sync::{Arc, Mutex};
 
@@ -12,6 +10,7 @@ use glass_core::{GlassError, Result};
 use serde_json::{json, Value};
 
 use crate::adb::Adb;
+use crate::conn::Conn;
 
 /// One absolute-display point in a pointer path (the agent's gesture element).
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
@@ -19,87 +18,6 @@ pub struct Pt {
     pub x: i32,
     pub y: i32,
     pub t_ms: u64,
-}
-
-/// The protocol version this client speaks (must match the agent's hello `proto`).
-const PROTO: i64 = 1;
-
-/// A live connection to the agent: a framed line reader/writer + a monotonic id.
-struct Conn {
-    writer: TcpStream,
-    reader: BufReader<TcpStream>,
-    next_id: i64,
-}
-
-impl Conn {
-    /// Connect to `127.0.0.1:port`, read + version-check the hello banner.
-    fn open(port: u16) -> Result<Conn> {
-        let stream = TcpStream::connect(("127.0.0.1", port))
-            .map_err(|e| GlassError::Backend(format!("agent connect :{port}: {e}")))?;
-        // Set read/write timeouts so a stalled agent surfaces as a transport error (which
-        // the existing reconnect path handles) rather than hanging the MCP thread forever.
-        let to = std::time::Duration::from_secs(30);
-        stream.set_read_timeout(Some(to)).ok();
-        stream.set_write_timeout(Some(to)).ok();
-        let reader = BufReader::new(
-            stream.try_clone().map_err(|e| GlassError::Backend(format!("agent clone: {e}")))?,
-        );
-        let mut c = Conn { writer: stream, reader, next_id: 1 };
-        let hello = c.read_line()?;
-        let v: Value = serde_json::from_str(&hello)
-            .map_err(|e| GlassError::Backend(format!("agent hello parse: {e}")))?;
-        let proto = v.get("hello").and_then(|h| h.get("proto")).and_then(Value::as_i64);
-        if proto != Some(PROTO) {
-            return Err(GlassError::Backend(format!(
-                "agent protocol mismatch: got {proto:?}, want {PROTO}"
-            )));
-        }
-        Ok(c)
-    }
-
-    fn read_line(&mut self) -> Result<String> {
-        let mut line = String::new();
-        let n = self
-            .reader
-            .read_line(&mut line)
-            .map_err(|e| GlassError::Backend(format!("agent read: {e}")))?;
-        if n == 0 {
-            return Err(GlassError::Backend("agent closed the connection".into()));
-        }
-        Ok(line.trim_end().to_string())
-    }
-
-    /// Send one request object (an `id` is injected) and return the response `Value`.
-    /// Returns `(result, io_error)` — `io_error` is true when the failure was a
-    /// transport I/O error (dropped connection) rather than a protocol-level error.
-    fn call(&mut self, mut req: Value) -> std::result::Result<Value, (GlassError, bool)> {
-        let id = self.next_id;
-        self.next_id += 1;
-        req["id"] = json!(id);
-        let mut line = serde_json::to_string(&req).expect("serialize request");
-        line.push('\n');
-        self.writer
-            .write_all(line.as_bytes())
-            .and_then(|_| self.writer.flush())
-            .map_err(|e| (GlassError::Backend(format!("agent write: {e}")), true))?;
-        let resp_line = self.read_line().map_err(|e| (e, true))?;
-        let resp: Value = serde_json::from_str(&resp_line)
-            .map_err(|e| (GlassError::Backend(format!("agent resp parse: {e}")), false))?;
-        if resp.get("id").and_then(Value::as_i64) != Some(id) {
-            return Err((
-                GlassError::Backend(format!(
-                    "agent response id mismatch (got {:?}, want {id})",
-                    resp.get("id")
-                )),
-                false,
-            ));
-        }
-        if resp.get("ok").and_then(Value::as_bool) != Some(true) {
-            let err = resp.get("error").and_then(Value::as_str).unwrap_or("agent error");
-            return Err((GlassError::Backend(format!("agent: {err}")), false));
-        }
-        Ok(resp)
-    }
 }
 
 /// Request/response client to the agent. `connect` reconnects on a dropped socket once.
